Encapsulate reference path generation
