+++ terrain_abort_planner/src/run_dynamic_rallypoints.cpp
@@ -55,6 +55,7 @@
 #include "terrain_navigation_ros/visualization.h"
 #include "terrain_planner/common.h"
 #include "terrain_planner/terrain_ompl_rrt.h"
+#include "terrain_abort_planner/mission_io.h"
 
 #include <nlohmann/json.hpp>
 using json = nlohmann::json;
@@ -219,25 +220,6 @@ void publishGridMap(const ros::Publisher& pub, const grid_map::GridMap& map) {
 
 double mod2pi(double x) { return x - 2 * M_PI * floor(x * (0.5 / M_PI)); }
 
-Eigen::Vector4d rpy2quaternion(double roll, double pitch, double yaw) {
-  double cy = std::cos(yaw * 0.5);
-  double sy = std::sin(yaw * 0.5);
-  double cp = std::cos(pitch * 0.5);
-  double sp = std::sin(pitch * 0.5);
-  double cr = std::cos(roll * 0.5);
-  double sr = std::sin(roll * 0.5);
-
-  Eigen::Vector4d q;
-  q(0) = cr * cp * cy + sr * sp * sy;
-  q(1) = sr * cp * cy - cr * sp * sy;
-  q(2) = cr * sp * cy + sr * cp * sy;
-  q(3) = cr * cp * sy - sr * sp * cy;
-
-  q.normalize();
-
-  return q;
-}
-
 PathSegment generateArcTrajectory(Eigen::Vector3d rate, const double horizon, Eigen::Vector3d current_pos,
                                   Eigen::Vector3d current_vel, const double dt = 0.1) {
   PathSegment trajectory;
@@ -275,83 +257,6 @@ PathSegment generateArcTrajectory(Eigen::Vector3d rate, const double horizon, Ei
 
     time = time + dt;
   }
-  return trajectory;
-}
-
-PathSegment getLoiterPath(Eigen::Vector3d end_position, Eigen::Vector3d end_velocity, Eigen::Vector3d center_pos) {
-  Eigen::Vector3d radial_vector = (end_position - center_pos);
-  radial_vector(2) = 0.0;  // Only consider horizontal loiters
-  Eigen::Vector3d emergency_rates =
-      20.0 * end_velocity.normalized().cross(radial_vector.normalized()) / radial_vector.norm();
-  double horizon = 2 * M_PI / std::abs(emergency_rates(2));
-  // Append a loiter at the end of the planned path
-  PathSegment loiter_trajectory = generateArcTrajectory(emergency_rates, horizon, end_position, end_velocity);
-  return loiter_trajectory;
-}
-
-PathSegment generateTrajectory(const double curvature, const double length, Eigen::Vector3d segment_start,
-                               Eigen::Vector3d segment_end, Eigen::Vector3d current_vel, const double dt) {
-  PathSegment trajectory;
-  trajectory.states.clear();
-  double progress = 0.0;
-  double flightpath_angle = std::sin((segment_end(2) - segment_start(2)) / length);
-  trajectory.flightpath_angle = flightpath_angle;
-  /// TODO: Fix sign conventions for curvature
-  trajectory.curvature = curvature;
-  trajectory.dt = dt;
-  // if (std::abs(curvature) < 0.0001) {
-  //   curvature > 0.0 ? trajectory.curvature = 0.0001 : trajectory.curvature = -0.0001;
-  // }
-
-  double current_yaw = std::atan2(-1.0 * current_vel(1), current_vel(0));
-  Eigen::Vector2d arc_center;
-  if (std::abs(curvature) > 0.001) {
-    arc_center = PathSegment::getArcCenter(trajectory.curvature, segment_start.head(2), current_vel.head(2),
-                                           segment_end.head(2));
-    double radial_yaw = std::atan2(-segment_start(1) + arc_center(1), segment_start(0) - arc_center(0));
-    current_yaw = curvature > 0.0 ? radial_yaw - 0.5 * M_PI : radial_yaw + 0.5 * M_PI;
-  }
-
-  for (int i = 0; i < std::max(1.0, length / dt); i++) {
-    State state_vector;
-    double yaw = -trajectory.curvature * progress + current_yaw;
-    if (std::abs(curvature) < 0.001) {
-      Eigen::Vector3d tangent = (segment_end - segment_start).normalized();
-      Eigen::Vector3d pos = progress * tangent + segment_start;
-
-      Eigen::Vector3d vel = current_vel;
-      const double roll = std::atan(trajectory.curvature / 9.81);
-      const double pitch = flightpath_angle;
-      Eigen::Vector4d att = rpy2quaternion(roll, -pitch, -yaw);  // TODO: why the hell do you need to reverse signs?
-
-      state_vector.position = pos;
-      state_vector.velocity = vel;
-      state_vector.attitude = att;
-    } else {
-      Eigen::Vector3d pos = (-1.0 / trajectory.curvature) * Eigen::Vector3d(std::sin(yaw) - std::sin(current_yaw),
-                                                                            std::cos(yaw) - std::cos(current_yaw), 0) +
-                            Eigen::Vector3d(0, 0, progress * std::sin(flightpath_angle)) + segment_start;
-
-      Eigen::Vector3d vel = Eigen::Vector3d(std::cos(flightpath_angle) * std::cos(yaw),
-                                            -std::cos(flightpath_angle) * std::sin(yaw), std::sin(flightpath_angle));
-      const double roll = std::atan(trajectory.curvature / 9.81);
-      const double pitch = flightpath_angle;
-      Eigen::Vector4d att = rpy2quaternion(roll, -pitch, -yaw);  // TODO: why the hell do you need to reverse signs?
-
-      state_vector.position = pos;
-      state_vector.velocity = vel;
-      state_vector.attitude = att;
-    }
-    trajectory.states.push_back(state_vector);
-
-    progress = progress + dt;
-  }
-  double end_yaw = -trajectory.curvature * length + current_yaw;
-  State end_state_vector;
-  end_state_vector.position = segment_end;
-  end_state_vector.velocity = Eigen::Vector3d(std::cos(end_yaw), -std::sin(end_yaw), 0.0);
-  end_state_vector.attitude = Eigen::Vector4d(1.0, 0.0, 0.0, 0.0);
-  trajectory.states.push_back(end_state_vector);
   return trajectory;
 }
 
@@ -443,47 +348,7 @@ int main(int argc, char** argv) {
     }
   }
   /// TODO: Generate Dubins path segments from waypoint list
-  Path mission_path;
-  double previous_distance_to_tangent = 0.0;
-  for (int waypoint_id = 1; waypoint_id < waypoint_list.size() - 1; waypoint_id++) {
-    const Eigen::Vector3d current_vertex = waypoint_list[waypoint_id];
-    const Eigen::Vector3d current_edge = current_vertex - waypoint_list[waypoint_id - 1];
-    const double current_edge_length = current_edge.norm();
-    // if (current_edge_length < 0.01) continue; //Same position
-    const Eigen::Vector3d current_tangent = current_edge / current_edge_length;
-    const double current_course = std::atan2(current_tangent[1], current_tangent[0]);
-
-    Eigen::Vector3d next_edge = waypoint_list[waypoint_id + 1] - current_vertex;
-    double next_edge_length = next_edge.norm();
-    Eigen::Vector3d next_tangent = next_edge / next_edge_length;
-    double next_course = std::atan2(next_tangent[1], next_tangent[0]);
-
-    double turn_angle = M_PI - std::abs(next_course - current_course);
-    double distance_to_tangent = radius / std::abs(std::tan(0.5 * turn_angle));
-    std::cout << "turn_angle: " << turn_angle / M_PI << " pi" << std::endl;
-    std::cout << "  - current_edge_length: " << current_edge_length << std::endl;
-    std::cout << "  - next_edge_length: " << next_edge_length << std::endl;
-    std::cout << "  - distance_to_tangent: " << distance_to_tangent << std::endl;
-
-    /// TODO: Arc segment
-    Eigen::Vector3d arc_segment_start = current_vertex - distance_to_tangent * current_tangent;
-    Eigen::Vector3d arc_segment_end = current_vertex + distance_to_tangent * next_tangent;
-    double arc_segment_length = std::abs(turn_angle) * radius;
-    double curvature = (next_course - current_course) > 0.0 ? 1.0 / radius : -1.0 / radius;
-    auto arc_segment =
-        generateTrajectory(curvature, arc_segment_length, arc_segment_start, arc_segment_end, current_tangent, 0.1);
-    /// TODO: Line segment
-    Eigen::Vector3d line_segment_start =
-        waypoint_list[waypoint_id - 1] + current_tangent * previous_distance_to_tangent;
-    Eigen::Vector3d line_segment_end = arc_segment_start;
-    double line_segment_length = current_edge_length - previous_distance_to_tangent - distance_to_tangent;
-    auto line_segment =
-        generateTrajectory(0.0, line_segment_length, line_segment_start, line_segment_end, current_tangent, 0.1);
-    previous_distance_to_tangent = distance_to_tangent;
-
-    mission_path.appendSegment(line_segment);
-    mission_path.appendSegment(arc_segment);
-  }
+  Path mission_path = generatePathFromWaypoints(waypoint_list, radius);
 
   /// Parse rally points from mission file
   std::vector<Eigen::Vector3d> rallypoint_list;
